Edited wiki page through web user interface.
I think use table is a good way there to clarify the things. Don't hesitate to go back If you think it's not good. :)

Thank for this page!

PS: I don't have time now but I think an exemple part (with real command line and explanations) could be good! (To set plugin for exemple) :)